Token Login and Register
Completed app routes to register users:

Passwords stored in the database are hashed using the bcrypt library for security.

Each account has a creation date stored
============================
Created a model for posts where each post stores a unique id, the userid of the person who created it, the description, and the text of the post.
============================
Created a token model that stores the id associated with the token, the refresh token, and dates of when it was created and when it should expired (have not implemented yet).
============================
Fully completed routes for http requests to get posts by userid, create new access tokens using a created refresh token, generate both refresh tokens and access tokens through login, and removing tokens from the database through a logout feature