Updated index.js and schema
Added a get request where you can get an article by its id

Updated /articles request by finding all the articles without needing an id.

Created a new http request that gets all of the articles that do not include the user that is logged in

articles now are created with a unique generated id for React routes.