Created articles page
Created a page where users can see articles they want to view.

Article Component is the article thumbnail where users can read the title and a small portion of the description of the article.

If the user clicks on the thumbnail, the project redirects them to a dynamic page where the articles can show the full description.