Create Article Component
Users are redirected to the "create-article" page where they can input their information that they want to be made as an article.

After pressing "submit article", the user is redirected to the articles page. Since they are the article of the new article, it will not be in their feed, but other users can view it.

Completed formatting for CreatePost and CreateArticle