API integration, custom waypoints, and booking
Integrated "IsItWater.com" API to detect if the coordinates of the middle between the user and the mentor is on water or land to calculate cafeterias.

Integrated Geoapify's reverse geocode API that converts coordinates into an address to use.

Integrated Geoapify's places API to look for cafeterias in the address area of the middlepoint. Each cafe displays the name, website, and directions.

Added a booking button where mentors can attach their booking link so users can get redirected to "Setmore" to book an appointment.