Reset Password Back End
Installed nodemailer to send emails through the back end.

Created a put request to change the new password while encrypting it to the database.

create a delete request where it verifies the code sent to the user with the code in the database. It checks if it is valid and if it has expired or not. If the code is valid, it will delete it from the database.

Created a post request to send an email to the user with their verification code. The program generates a 6 digit code and adds an expiration date set for five minutes. The information is saved in the database for comparison.