Front End
Added the app route for reset password

Added an anchor element in the login page to redirect the user to the reset password page

Program guides user to enter their email to sends them a verification code.

The user verifies the verification code and are prompted to change their password.

Included user error messages to display in the UI.