TC2 Change
Solved the approach if the middlepoint between two users are on water by using a route API to get the coordinates throughout the route.

Created a function to find the middlepoint coordinates by finding which one has the closest distance to the coordinates that is on water.

Added comments to make the approach understandable