Filter Feature and CSS Changes
Added a filter feature where users can filter posts by who they are following, oldest, or newest posts.

Moved the comments inside of the posts box for a new design.

Added animations to posts when loading or hovering on it.

The company logo image has an animation when you hover on it.